Added firmware files removing one change done to FW
Fimrware removes one of the two changes done to the software to implement a hight QoS, to see if it is the general QoS increase causing problem, or just the setting of an additional flag in the creating of an instance of the CloudManager. Version is also used to see if we can clock a significant improvent in the sending of the firmware-update to the display, as well as the sending of Frimware Status prior to starting update, and the sending of a new firmware version after the update has been successfully performed.